Urql graphql client #8 split to load and display component
